Links Text einblenden für Anzahl der Traces, +Sortierung + etwas aufräumen
